requirements.txt
boto3==1.34.100
langchain==0.1.17
mtranslate==1.8
streamlit==1.34.0
sentence_transformers
faiss-gpu
